+++ packages/website/ts/pages/about/team.tsx
@@ -39,11 +39,6 @@ const team: TeamMember[] = [
         title: 'Director of Operations',
     },
     {
-        imageUrl: '/images/team/leol.jpg',
-        name: 'Leonid Logvinov',
-        title: 'Engineer',
-    },
-    {
         imageUrl: '/images/team/benb.jpg',
         name: 'Ben Burns',
         title: 'Designer',
@@ -64,6 +59,11 @@ const team: TeamMember[] = [
         title: 'Ecosystem Engineer',
     },
     {
+        imageUrl: '/images/team/greg.jpg',
+        name: 'Greg Hysen',
+        title: 'Blockchain Engineer',
+    },
+    {
         imageUrl: '/images/team/blake.jpg',
         name: 'Blake Henderson',
         title: 'Ecosystem Programs Lead',
@@ -72,11 +72,6 @@ const team: TeamMember[] = [
         imageUrl: '/images/team/zack.jpg',
         name: 'Zack Skelly',
         title: 'Lead Recruiter',
-    },
-    {
-        imageUrl: '/images/team/greg.jpg',
-        name: 'Greg Hysen',
-        title: 'Blockchain Engineer',
     },
     {
         imageUrl: '/images/team/remcoB.jpg',
@@ -182,6 +177,16 @@ const team: TeamMember[] = [
         imageUrl: '/images/team/masonl.jpg',
         name: 'Mason Liang',
         title: 'Research Engineer',
+    },
+    {
+        imageUrl: '/images/team/patryka.jpg',
+        name: 'Patryk Adaś',
+        title: 'Designer',
+    },
+    {
+        imageUrl: '/images/team/alexk.jpg',
+        name: 'Alex Kroeger',
+        title: 'Data Scientist',
     },
 ];