+++ python-packages/contract_addresses/src/zero_ex/contract_addresses/__init__.py
@@ -75,7 +75,7 @@ NETWORK_TO_ADDRESSES: Dict[NetworkId, ContractAddresses] = {
         forwarder="0x5468a1dc173652ee28d249c271fa9933144746b1",
         order_validator="0x9463e518dea6810309563c81d5266c1b1d149138",
         coordinator_registry="0x45797531b873fd5e519477a070a955764c1a5b07",
-        coordinator="0x24675738816c87ad37e712cc24f309a0c906187f",
+        coordinator="0x25aae5b981ce6683cc5aeea1855d927e0b59066f",
     ),
     NetworkId.ROPSTEN: ContractAddresses(
         erc20_proxy="0xb1408f4c245a23c31b98d2c626777d4c0d766caa",
@@ -87,7 +87,7 @@ NETWORK_TO_ADDRESSES: Dict[NetworkId, ContractAddresses] = {
         forwarder="0x2240dab907db71e64d3e0dba4800c83b5c502d4e",
         order_validator="0x90431a90516ab49af23a0530e04e8c7836e7122f",
         coordinator_registry="0x403cc23e88c17c4652fb904784d1af640a6722d9",
-        coordinator="0xc442300dcb4df1ff1db0173e77556dc559de6006",
+        coordinator="0x25aae5b981ce6683cc5aeea1855d927e0b59066f",
     ),
     NetworkId.RINKEBY: ContractAddresses(
         exchange="0xbce0b5f6eb618c565c3e5f5cd69652bbc279f44e",
@@ -99,7 +99,7 @@ NETWORK_TO_ADDRESSES: Dict[NetworkId, ContractAddresses] = {
         forwarder="0x2d40589abbdee84961f3a7656b9af7adb0ee5ab4",
         order_validator="0x0c5173a51e26b29d6126c686756fb9fbef71f762",
         coordinator_registry="0x1084b6a398e47907bae43fec3ff4b677db6e4fee",
-        coordinator="0xc05e6c01b83eeedecac2f7bab9367da4d4c108df",
+        coordinator="0x25aae5b981ce6683cc5aeea1855d927e0b59066f",
     ),
     NetworkId.KOVAN: ContractAddresses(
         erc20_proxy="0xf1ec01d6236d3cd881a0bf0130ea25fe4234003e",
@@ -111,7 +111,7 @@ NETWORK_TO_ADDRESSES: Dict[NetworkId, ContractAddresses] = {
         forwarder="0x17992e4ffb22730138e4b62aaa6367fa9d3699a6",
         order_validator="0xb389da3d204b412df2f75c6afb3d0a7ce0bc283d",
         coordinator_registry="0x09fb99968c016a3ff537bf58fb3d9fe55a7975d5",
-        coordinator="0x04b2b090bad68b254881d7eb645a258ce66cc998",
+        coordinator="0x25aae5b981ce6683cc5aeea1855d927e0b59066f",
     ),
     NetworkId.GANACHE: ContractAddresses(
         exchange="0x48bacb9266a570d521063ef5dd96e61686dbe788",